LPS-196240 Move ObjectActionThreadLocal to object-api to use in in test
