+++ modules/apps/object/object-api/bnd.bnd
@@ -5,6 +5,7 @@ Export-Package:\
 	com.liferay.object.action.engine,\
 	com.liferay.object.action.executor,\
 	com.liferay.object.action.trigger,\
+	com.liferay.object.action.util,\
 	com.liferay.object.configuration,\
 	com.liferay.object.constants,\
 	com.liferay.object.definition.notification.term.util,\
