+++ modules/apps/object/object-api/src/main/java/com/liferay/object/action/util/ObjectActionThreadLocal.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.internal.action.util;
+package com.liferay.object.action.util;
 
 import com.liferay.petra.lang.CentralizedThreadLocal;
